Rename package common -> base
